Move RumConstants to the agent
