+++ android-agent/src/main/java/io/opentelemetry/android/RuntimeDetailsExtractor.java
@@ -5,9 +5,9 @@
 
 package io.opentelemetry.android;
 
-import static io.opentelemetry.android.common.RumConstants.BATTERY_PERCENT_KEY;
-import static io.opentelemetry.android.common.RumConstants.HEAP_FREE_KEY;
-import static io.opentelemetry.android.common.RumConstants.STORAGE_SPACE_FREE_KEY;
+import static io.opentelemetry.android.internal.tools.RumConstants.BATTERY_PERCENT_KEY;
+import static io.opentelemetry.android.internal.tools.RumConstants.HEAP_FREE_KEY;
+import static io.opentelemetry.android.internal.tools.RumConstants.STORAGE_SPACE_FREE_KEY;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
