+++ android-agent/src/main/java/io/opentelemetry/android/features/diskbuffering/scheduler/DefaultExportScheduler.kt
@@ -6,9 +6,9 @@
 package io.opentelemetry.android.features.diskbuffering.scheduler
 
 import android.util.Log
-import io.opentelemetry.android.common.RumConstants.OTEL_RUM_LOG_TAG
 import io.opentelemetry.android.features.diskbuffering.SignalFromDiskExporter
 import io.opentelemetry.android.internal.services.periodicwork.PeriodicRunnable
+import io.opentelemetry.android.internal.tools.RumConstants.OTEL_RUM_LOG_TAG
 import java.io.IOException
 import java.util.concurrent.TimeUnit
