+++ android-agent/src/main/java/io/opentelemetry/android/internal/features/persistence/DiskManager.java
@@ -6,11 +6,11 @@
 package io.opentelemetry.android.internal.features.persistence;
 
 import android.util.Log;
-import io.opentelemetry.android.common.RumConstants;
 import io.opentelemetry.android.features.diskbuffering.DiskBufferingConfiguration;
 import io.opentelemetry.android.internal.services.CacheStorageService;
 import io.opentelemetry.android.internal.services.PreferencesService;
 import io.opentelemetry.android.internal.services.ServiceManager;
+import io.opentelemetry.android.internal.tools.RumConstants;
 import java.io.File;
 import java.io.IOException;
