+++ android-agent/src/test/java/io/opentelemetry/android/OpenTelemetryRumBuilderTest.java
@@ -5,8 +5,8 @@
 
 package io.opentelemetry.android;
 
-import static io.opentelemetry.android.common.RumConstants.SCREEN_NAME_KEY;
-import static io.opentelemetry.android.common.RumConstants.SESSION_ID_KEY;
+import static io.opentelemetry.android.internal.tools.RumConstants.SCREEN_NAME_KEY;
+import static io.opentelemetry.android.internal.tools.RumConstants.SESSION_ID_KEY;
 import static io.opentelemetry.sdk.testing.assertj.OpenTelemetryAssertions.assertThat;
 import static io.opentelemetry.sdk.testing.assertj.OpenTelemetryAssertions.equalTo;
 import static org.awaitility.Awaitility.await;
