+++ android-agent/src/test/java/io/opentelemetry/android/SessionIdSpanAppenderTest.java
@@ -5,7 +5,7 @@
 
 package io.opentelemetry.android;
 
-import static io.opentelemetry.android.common.RumConstants.SESSION_ID_KEY;
+import static io.opentelemetry.android.internal.tools.RumConstants.SESSION_ID_KEY;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.verify;
